+++ for_src/diag_balance_io.F90
@@ -7,7 +7,7 @@ subroutine init_diag_balance
  include "netcdf.inc"
  include "mpif.h"
  integer :: ncid,iret,xdim,ydim,zdim,tdim,id,i,j,k,nc_mode
- real*8 :: x(nx),y(ny),z(nz)
+ real(real_type) :: x(nx),y(ny),z(nz)
  
  call allocate_module_diag_balance
  
@@ -140,7 +140,7 @@ subroutine write_diag_balance
  include "netcdf.inc"
  include "mpif.h"
  integer :: ncid,iret,id,nc_mode,start(4),count(4),ilen,k
- real*8 :: aloc(is_pe:ie_pe,js_pe:je_pe,ks_pe:ke_pe)
+ real(real_type) :: aloc(is_pe:ie_pe,js_pe:je_pe,ks_pe:ke_pe)
  
  if (my_pe==0)  print*,'writing to file diag_balance.cdf'
 
@@ -251,7 +251,7 @@ subroutine init_diag_balance
  implicit none
  include "netcdf.inc"
  integer :: ncid,iret,i,xdim,ydim,zdim,timeid,timedim,id
- real*8 :: x(nx),y(ny),z(nz)
+ real(real_type) :: x(nx),y(ny),z(nz)
  
  call allocate_module_diag_balance
  
@@ -515,7 +515,7 @@ subroutine write_diag_balance
  integer :: tdimid,ilen,timeid,id
  integer :: tag=1,ist(3),isz(3),ien(3)
  integer, dimension(MPI_STATUS_SIZE) :: Status
- real*8, allocatable :: a(:,:,:)
+ real(real_type), allocatable :: a(:,:,:)
 
  if (my_pe==0) then
    print*,'writing to file diag_balance.cdf at t=',time
